Move git and logger to utils
