+++ kea-vscode/src/utils/logger.ts
